Safari-609.1.11.1/LayoutTests/media/modern-media-controls/css/transformed-media-expected.txt

PASS mediaControls.style.width is "560px"
PASS mediaControls.style.height is "260px"
PASS successfullyParsed is true

TEST COMPLETE
